requirements.txt
fpdf
pandas
numpy
matplotlib
seaborn
emoji
Flask-Images